feat. 컴포넌트 내부에서 관리되는 상수 > 상수파일로 분리
