+++ src/components/report/ChatDetail.tsx
@@ -1,4 +1,5 @@
 import { BarChart, Bubble, ReportTitle } from 'components';
+import { KEYWORD_NAME_LIST } from 'shared/constants/constants';
 import useReport from 'shared/query/useReport';
 import isMobile from 'shared/utils/deviceDetector';
 import * as t from './chatDetail.style';
@@ -14,7 +15,13 @@ export default function ChatDetail() {
           <h1>
             대화에서 많은 선택을 받은 대화 주제는{' '}
             <span>
-              '{KEYWORD_LIST[chatData?.data.keyword_score[0].keyword_id - 1]}'
+              '
+              {
+                KEYWORD_NAME_LIST[
+                  chatData?.data.keyword_score[0].keyword_id - 1
+                ]
+              }
+              '
             </span>{' '}
             이에요.
           </h1>
@@ -23,7 +30,7 @@ export default function ChatDetail() {
               ({ keyword_id, keyword_count }) => (
                 <BarChart
                   key={keyword_id}
-                  text={KEYWORD_LIST[keyword_id - 1]}
+                  text={KEYWORD_NAME_LIST[keyword_id - 1]}
                   value={keyword_count}
                 />
               )
@@ -62,16 +69,3 @@ export default function ChatDetail() {
     </t.Container>
   );
 }
-
-const KEYWORD_LIST = [
-  '일상',
-  '관계',
-  '나',
-  '휴식',
-  '미래/성장',
-  '여행',
-  '팀',
-  '커리어',
-  '사랑',
-  '일',
-];
